Fix KEYRESET routine
KEYRESET was causing random keys to be detected as STOP key because the KEYBOARD ROW was left in a random state. It now sets it to ROW 0.  NOTE: The STOP key location is hardcoded in the KERNAL for B and N keyboards. Other keyboards (such as DIN) may have STOP in a different location and may require a patch in the keyboard scanning routine.