Add comments to B keyboard scanner
Trying to understand REPEAT key which is fairly useless on the B keyboard.  Thinking it might be used to generate the graphics characters that can no longer be generated due to SHIFTED top-line symbols. On the N keyboard the top-row keys are UN-SHIFTED symbols and SHIFT produces the graphic characters.