Re-fixed the KEYRESET routine
I had forgotten that the keyboard is scanned from 0 to 9 but that the keyboard matrix is stored in reverse order. In KEYRESET I was incorrectly setting the ROW to 0 instead of 9. Strangely for the C64 I set it correctly, but in reality there is no need to handle them differently.